+++ CHANGES.md
@@ -1,6 +1,7 @@
 unreleased
 ----------
 
+- Turned off address space layout randomization on Windows 1809. (@mtelvers #196)
 - Add Ubuntu 23.10. (@MisterDA #189)
 - Deprecate Ubuntu 22.10 it is now EOL. (@tmcgilchrist #184).
 - Support `--start-interval` in `HEALTCHECK` Dockerfile instruction.